feat: automatic reconnections
This commit introduces automatic reconnections. I've removed calls to `ConnectedBridges` as it's proven to be an unreliable way to check if a bridge is connected. Instead I list the bridges and try to manually reconnect. If the reconnect is successful then I have a bridge connection, if it fails, I know for certian I'm in a disconnected state. This seems to mitigate a lot of problems with automatic reconnections. I still don't rely on the connection status stream, but I think this is simplest for now.