feat: basic logging
This commit introduces basic logging of the requests and responses from any commands the renderer process sends to the main process. The logs live here for windows:

```
%USERPROFILE%\AppData\Roaming\{app name}\logs\{process type}.log
```

In dev mode they print to the console and it's been working OK. We'll have to see how fast they fill up later.