feat: add version
We need a way to check to make sure the version is correct in the MyRC+S app. This commit adds a little version tag based on the version contained in the `package.json` file. For that reason it probably relies on some of the work @suivo2001 is doing with the git modules to autoupdate the version