feat: disable auto update
This commit introduces a new config flag to prevent auto update. This is useful in the cases where users are stuck on old versions of the OMNI server. Auto update for users stuck with old versions of OMNI would update the client application so that it would error when talking with OMNI